+++ crates/cpu/src/opcodes.rs
@@ -1,13 +1,13 @@
 mod arithmetic;
 mod bits8;
 mod cb_operation;
-mod consts;
 mod data;
+mod load;
 mod load16b;
 mod rotate;
 
 pub use arithmetic::Arithmetic;
-pub use bits8::Load;
+pub use load::Load;
 pub use data::Data;
 pub use load16b::Load16b;
 pub use rotate::Rotate;
