--- crates/cpu/src/opcodes/bits8.rs
@@ -1,3 +1,0 @@
-mod load;
-
-pub use load::Load;
